# server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api github Documentation",
    "description": "Github API",
    "host": "localhost:3001",
    "version": "1.0.0",
    "contact": {
      "email": "andressalhenriques@gmail.com"
    }
  },
  "paths": {
    "/repositories": {
      "post": {
        "tags": ["Repositories"],
        "summary": "Search repositories  by repository name",
        "description": "Search repositories  by repository name",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Repository not found"
          }
        }
      }
    },
    "/createBookmarksRepositories": {
      "post": {
        "tags": ["CreateBookmarksRepositories"],
        "summary": "Create repositories bookmarks",
        "description": "Create repositories bookmarks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of bookmarks"
          },
          "500": {
            "description": "not working"
          }
        }
      }
    },
    "/bookmarksRepositories": {
      "post": {
        "tags": ["Find all repositories bookmarked"],
        "summary": "Find all repositories bookmarked",
        "description": "Find all repositories bookmarked",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of bookmarks"
          },
          "500": {
            "description": "not working"
          }
        }
      }
    }
  }
}